Examples Update in will-change Article
Updated examples removing property names that won't be optimized and replacing them with ones that can be. For example, based on some feedback I got from David Baron, and I agree, the `animation` property does not hint the browser which properties will change. Instead, the properties that are going to change in the keyframes block should be used as values for `will-change`. I updated the CSS and JS accordingly. Same applies for `position` property: instead of using `position` which doesn't get any optimizations, I just used `top`, `left`, etc.